Fix runUnderRealCredentials failing to switch back to fake creds
When we reset the cred manager, it causes a switch to dummy creds due to the active has now been removed. This ends up getting cancelled by the explicit switch back to oldActive which is the same (DUMMY). So instead switch the active, before the removal of the real creds