Make the dotnet lambda image tests assert something, fix test data
The dockerfiles referenced Lambda handlers that did not eixst. This caused the tests to not actually run the Lambda handler but SAM would still be exit code 0 so tests would pass. Add an assert on the cred env vars to make sure handler is ran. Also make the handler correct in the dockerfiles