Add missing userDecision events from conflicting popup
1. When CodeWhisperer is invoked but IntelliSense popup is opened
after, we are not showing recommendations due to conflicting popups
in this case previously we rely on disposing the popup from the caller
processCodeWhispererUI(), but the popupListener to send userDecision
event has not yet been registered to the popup. So we send the event
manually just like the caret moving backward case.