Fix Enter action will add incorrect number of newlines when triggered inside braces
1. When enter action is triggered inside braces, an EnterBetweenBracesFinalHandler
will be invoked which is an EnterHandlerDelegate, this delegate will call
originalHandler.execute() to add an extra newline to correctly put the closing
brace in another new line. Inside CodeWhispererEnterHandler we need to call
originalHandler.execute() instead of super.executeWriteAction() to avoid
adding another(3rd) newline character in the document.
2. Since the originalHandler in EnterHandler is private not protected, need to
make the originalHander in CodeWhispererEnterHandler a private field.