Make CW auto-reconnect on startup
This fixes the issue when CW does not auto-reconnect on IDE startup.
CW will try to auto-connect when the connection(accessToken is expired)
It will first try to use the refresh token to refresh, if failed, it
will prompt a notification for people to re-auth(go through the browser
login flow again).
Also did some refactors to ensure:
1. CW actions are not performed when CW is logged out, or CW needs to re-auth,
for the case when CW needs to refresh, refresh the token and continue with
the CW action.
2. Reuse helpers for serveral parts for reconnect, refresh, showing expiry
notification and expiry check.