fix: [Amazon Q] handle exceptions in service request
Currently, most exceptions that occur when the amazon Q API is invoked
are suppressed and are not surfaced to the user (For ex,
`AccessDeniedException`).

The plugin proceeds like the call succeeded, and subsequently makes another
call to add a telemetry event that also fails due to an improperly formed
request (from not having a conversationId), and the user is shown this error
instead of the original error.

This change fixed that.