Make sandbox run configurations more representative of the published artifact
Previously, the sandbox tasks only ran a given submodule of the plugin.
This led to some scenarios where no issues were found when testing in the IDE sandbox, but were discovered once they were published in the marketplace.
Resolved by always running the sandbox on the artifact that will be published to the marketplace