/dev: include in progress state agent details in code generation (#4828)
Currently we display a long-running message "Generating code ... " while in code generation still IN_PROGRESS state. There is an often customer feedback based on this experience that its "stuck" or not "progressing".

Although the RTS and LLM are processing the request, user expects some feedback on whats is doing or if its actually stuck processing.