Fix UI freeze when trying to refresh tokens offline (#4880)
`CodeWhispererService#showRecommendationsInPopup` was previously running on EDT, which caused issues when we need to wait for the fairly expensive call chain for `promptReauth` if the connection is expired. This is very visible in the offline case.

Fix by moving all the operations onto a background worker and limiting invocations to one at a time. This caused some issues with the test suite since they assumed that `showRecommendationsInPopup` would block until background operations have initialized, which was worked around by passing along job tracking