Track SessionContext instead of InvocationContext
1. Since the shift from displaying one trigger at a time
to displaying multiple triggers at a time, use SessionContext
as the disposable for the session states.
2. Introduce the concept of "display session" which can show
multiple triggers at a time and sessionContext is to hold states
for that session.
3. ongoingRequests will be a map of <jobId, InvocationContext>
where InvocationContext is the state of each trigger.
4. Managed these states globally by dynamically adding Trigger states
into ongoingRequests and whenever the user made a decision(accept or
reject), conclude decisions for all the ongoing triggers and clear
ongoingRequest, so that a new display session can start later.