Add OpenTelemetry base classes (#4982)
The OpenTelemetry SDK can automatically attach the same trace ID to spans recorded under the same 'context'. This is somewhat challenging because the JetBrains platform frequently requires switching threads and coroutine scope contexts, which loses the `ThreadLocal` state needed to connect parent->child when spans are started.

This on its own is not too difficult, but becomes a significant ergonomic challenge when we also want to bring some type-safety to how span metadata is built. This PR allows spans generated in https://github.com/aws/aws-toolkit-common/pull/899 to automatically perform context propagation.

See internal documentation for details. A contributor-friendly version of the doc will be written after we validate this pattern with some initial metrics.