Using AB variation value as customization name when overridden (#5090)
When a customizationArnOverride is present from an AB feature response, the toolkit will currently default to displaying the name of the last selected customization by the user rather than a name for the customization that was overridden. With this change, we'll use the AB variation field to specify a name for the customizationArn override and use it for display.