telemetry(amazonq): Add diff telemetry for generated and accepted changes. (#5116)
The purpose of this change is to introduce telemetry for generated and accepted code within Feature Dev (`/dev`). This change performs a diff when code is generated or accepted, and records generated code. This change handles duplicate generation/acceptance within a given conversation by tracking an identifier of generated changes for a given file based on a hash of the before/after content.