fix(amazonq): Normalize generated file paths in feature dev. (#5130)
The agent may generate files with or without a leading slash in the generated file path. In JetBrains, this results in an exception thrown when attempting to write to the root directory on MacOS. (Various other failure modes may occur due to lack of normalization, depending on OS.)

This change normalized handling within /dev, to avoid breakage or other side effects of non-normalized file paths.