Fix JSR310 serialization error when preparing login panel
If a user is somehow in the situtation where they have an `sso-session` defined,
with a token in their SSO cache (either valid/invalid), then prepareBrowser fails
since it tries to serialize the state of `connectionSettings` without the jackson-jsr310 module.

This is not intended behavior, so fix by configuring this to be an ignored field.