Adds event listener for notifying UI that AB feature configurations have been resolved (#5308)
Adds a new event listener object and handler to allow the application to notify the UI when feature configs have been resolved. For now, this is just the highlight command as that is all the UI needs. In the future we can use this class to plumb through additional features.