config(amazonq): update service sdk model (#1986)
pre-requisite of #1984.

This PR updates the service-2.json file to align with the latest SDK model.

Modified unit tests as the SDK models for CreateCodeScanRequest and StartCodeAnalysisRequest are no longer identical, making equals() comparisons invalid.

Verified CodeScan functionality from a functional perspective to ensure the updated model works as expected.