fix(amazonq): use https as scheme for local webview resource resolver (#5489)
https://developer.mozilla.org/en-US/docs/Web/API/Clipboard
`navigator.clipboard` is only accessible from a "secure context", which JCEF does not think we are using,
because the URI authority is not localhost or https. Since we are hooking onto the CEF resource request handler,
it does not matter what we use for the scheme/authority as long as there is no conflict with real resources.

Previous implementation used a different load mechanism which somehow met the secure context criteria.