fix(amazonq): support displaying customizations across all profiles of an Idc connection (#5591)
Follow-up to #5575.

As mentioned in that PR:

Before: Customizations were bound to a specific IDC instance.
After: Customizations are now bound to a specific Q profile. An IDC instance can have multiple Q profiles.

In other words, each Q profile has access to its own set of customizations.

The product team has requested that we show all available customizations across profiles, rather than only those tied to the currently connected profile.

This means that when a user selects a customization that belongs to a different Q profile, the plugin should implicitly switch to that profile. This behavior helps reduce confusion and user churn, since users might not clearly understand which profile grants access to which customizations.