telemetry(amazonq): Add changed IDE diagnostics after user acceptance (#5613)
When user accepts suggestion from inline completion or chat, there can be a change in the current open editor's IDE diagnostics, this can be used as a measure for code suggestion quality. Ref https://github.com/aws/aws-toolkit-vscode/pull/7130.

This change is part of the server side workspace context.

To reduce the risk of large blast radius, this change is only applied for users who are in the experiment of server side project context (both treatment and control), which is the Amzn idc users. 

Verified backend request id 6faef702-938e-42eb-894a-957d9757186e