fix(amazonq): ChatCommunicationManager blocks messages to UI until ready (#5697)
sendContextCommands pulls in changes required for the context selection/provider prompts. The server makes this call as soon as it gets the config. There may be situations where context calls are made before we can display them in the browser.

Messages received by the client are sent to the browser through the chat manager and blocked until the ui is ready.