feat(amazonq): pass workspaceIdentifier when initializing AmazonQLanguageServer (#5725)
AmazonQ LSP needs an identifier for the IDE workspace, which should be stable and unique for each workspace, regardless of IDE restarts or system reboots.

Related `aws/language-server-runtimes` change:
- https://github.com/aws/language-server-runtimes/pull/497



Use `Project.getBasePath()` as such an identifier and pass it when initializing AmazonQ language server.

Similar change on `aws/aws-toolkit-vscode` side:
- https://github.com/aws/aws-toolkit-vscode/pull/7291