merge: support chat on remote connections (#6008)
* feat(amazonq): Add support for Amazon Q chat on remote 242+ (#4825)

* fix(amazonq): Bug Fix for "Q: Send to Prompt" doesn't work in split panes on local or remote (#5463)

* fix(amazonq): use https as scheme for local webview resource resolver (#5489)

https://developer.mozilla.org/en-US/docs/Web/API/Clipboard
`navigator.clipboard` is only accessible from a "secure context", which JCEF does not think we are using,
because the URI authority is not localhost or https. Since we are hooking onto the CEF resource request handler,
it does not matter what we use for the scheme/authority as long as there is no conflict with real resources.

Previous implementation used a different load mechanism which somehow met the secure context criteria.

* fix(amazonq): fix agentic asset loading in remote mode (#5734)

local uris cannot be loaded if the host is deemed non-remote

* fix(amazonq): fix code issues tab not appearing in remote (#5737)

on remote, each client gets its own instance of the tool window.

logic needs to handle this case, otherwise it will only make changes against the 'main' IDE,
and will not be visible to remote clients

* fix(amazonq): use progress bar decorator while loading q chat on remote (#5749)

animated icons do not show up on remote

* fix: provide loading text when chat assets are slow to load (#5973)

on remote, it can take 10+ seconds for chat to be visible.

additionally, delete some unneeded assets to slightly reduce load time

* feat(amazonq): limit remote to amzn-internal compute (#5993)

Per internal discussion, feature deemed not ready for external users due to https://youtrack.jetbrains.com/issue/IJPL-203169

* fix: delete changelog for internal feature (#6009)

* feat: show message to resize on remote chat (#6007)

---------
Co-authored-by: Lokesh <lodogga@amazon.com>